Chaskis/UnitTests/CoreTests/Handlers/Receive/ReceiveHandlerConfigTests.cs
//
//          Copyright Seth Hendrick 2016-2020.
// Distributed under the Boost Software License, Version 1.0.
//    (See accompanying file LICENSE_1_0.txt or copy at
//          http://www.boost.org/LICENSE_1_0.txt)
//

using Chaskis.Core;
using NUnit.Framework;
using SethCS.Exceptions;

namespace Chaskis.UnitTests.CoreTests.Handlers.Receive
{
    [TestFixture]
    public class ReceiveHandlerConfigTests
    {
        // ---------------- Tests ----------------

        /// <summary>
        /// Ensures the Validate() function works as expected.
        /// </summary>
        [Test]
        public void ValidateTest()
        {
            ReceiveHandlerConfig config = new ReceiveHandlerConfig();

            config.LineAction = null;
            Assert.Throws<ValidationException>( () => config.Validate() );

            config.LineAction = delegate ( ReceiveHandlerArgs args )
            {
            };

            Assert.DoesNotThrow( () => config.Validate() );
        }

        [Test]
        public void CloneTest()
        {
            ReceiveHandlerConfig config1 = new ReceiveHandlerConfig();
            ReceiveHandlerConfig clone = config1.Clone();

            Assert.AreNotSame( config1, clone );
        }
    }
}
